fix: 잠재적 결함 해결
